@@ -5,12 +5,12 @@ homepage: https://z.flutter.gallery/
 repository: https://github.com/jamesblasco/zflutter
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
-  quiver: ^2.1.3
+  quiver: 3.1.0
 
 dev_dependencies:
   flutter_test:
