@@ -4,22 +4,22 @@ import 'package:flutter/material.dart';
 import 'package:zflutter/zflutter.dart';
 
 class ZBox extends StatelessWidget {
-  final double width;
-  final double height;
-  final double depth;
+  final double? width;
+  final double? height;
+  final double? depth;
 
   final double stroke;
   final bool fill;
 
-  final Color color;
-  final bool visible;
+  final Color? color;
+  final bool? visible;
 
-  final Color frontColor;
-  final Color rearColor;
-  final Color leftColor;
-  final Color rightColor;
-  final Color topColor;
-  final Color bottomColor;
+  final Color? frontColor;
+  final Color? rearColor;
+  final Color? leftColor;
+  final Color? rightColor;
+  final Color? topColor;
+  final Color? bottomColor;
 
   ZBox({
     this.width,
@@ -38,22 +38,22 @@ class ZBox extends StatelessWidget {
   });
 
   Widget get frontFace => ZPositioned(
-        translate: ZVector.only(z: depth / 2),
+        translate: ZVector.only(z: depth! / 2),
         child: ZRect(
           color: frontColor ?? color,
           fill: fill,
           stroke: 1,
-          width: width,
-          height: height,
+          width: width!,
+          height: height!,
         ),
       );
 
   Widget get rearFace => ZPositioned(
-        translate: ZVector.only(z: -depth / 2),
+        translate: ZVector.only(z: -depth! / 2),
         rotate: ZVector.only(y: tau / 2),
         child: ZRect(
-          width: width,
-          height: height,
+          width: width!,
+          height: height!,
           color: rearColor ?? color,
           fill: fill,
           stroke: 1,
@@ -61,11 +61,11 @@ class ZBox extends StatelessWidget {
       );
 
   Widget get leftFace => ZPositioned(
-        translate: ZVector.only(x: -width / 2),
+        translate: ZVector.only(x: -width! / 2),
         rotate: ZVector.only(y: -tau / 4),
         child: ZRect(
-          width: depth,
-          height: height,
+          width: depth!,
+          height: height!,
           stroke: 1,
           color: leftColor ?? color,
           fill: fill,
@@ -73,38 +73,38 @@ class ZBox extends StatelessWidget {
       );
 
   Widget get rightFace => ZPositioned(
-        translate: ZVector.only(x: width / 2),
+        translate: ZVector.only(x: width! / 2),
         rotate: ZVector.only(y: tau / 4),
         child: ZRect(
-          width: depth,
+          width: depth!,
           color: rightColor ?? color,
-          height: height,
+          height: height!,
           stroke: 1,
           fill: fill,
         ),
       );
 
   Widget get topFace => ZPositioned(
-        translate: ZVector.only(y: -height / 2),
+        translate: ZVector.only(y: -height! / 2),
         rotate: ZVector.only(x: -tau / 4),
         child: ZRect(
-          width: width,
+          width: width!,
           color: topColor ?? color,
-          height: depth,
+          height: depth!,
           stroke: 1,
           fill: fill,
         ),
       );
 
   Widget get bottomFace => ZPositioned(
-        translate: ZVector.only(y: height / 2),
+        translate: ZVector.only(y: height! / 2),
         rotate: ZVector.only(x: tau / 4),
         child: ZRect(
-          width: width,
+          width: width!,
           color: bottomColor ?? color,
           stroke: 1,
           fill: fill,
-          height: depth,
+          height: depth!,
         ),
       );
 
