@@ -7,9 +7,9 @@ class ZIllustration extends ZMultiChildWidget {
   final double zoom;
 
   ZIllustration({
-    List<Widget> children,
+    required List<Widget> children,
     this.zoom = 1,
-  })  : assert(zoom != null && zoom >= 0),
+  })  : assert(zoom >= 0),
         super(children: children);
 
   @override
@@ -25,20 +25,20 @@ class ZIllustration extends ZMultiChildWidget {
 }
 
 class RenderZIllustration extends RenderZMultiChildBox {
-  double _zoom = 1;
+  double? _zoom = 1;
 
-  double get zoom => _zoom;
+  double? get zoom => _zoom;
 
-  set zoom(double value) {
-    assert(_zoom != null && _zoom >= 0);
+  set zoom(double? value) {
+    assert(_zoom != null && _zoom! >= 0);
     if (_zoom == value) return;
     _zoom = value;
     markNeedsPaint();
   }
 
   RenderZIllustration({
-    double zoom,
-    List<RenderZBox> children,
+    double? zoom,
+    List<RenderZBox>? children,
   })  : assert(zoom != null && zoom >= 0),
         _zoom = zoom,
         super(children: children, sortMode: SortMode.update);
@@ -58,7 +58,7 @@ class RenderZIllustration extends RenderZMultiChildBox {
 
   // TODO: Work on hitTest
   @override
-  bool hitTest(BoxHitTestResult result, {Offset position}) {
+  bool hitTest(BoxHitTestResult result, {required Offset position}) {
     assert(() {
       if (!hasSize) {
         if (debugNeedsLayout) {
